wtf why don't you push

+++ Library/core/src/main/java/hu/uni/miskolc/iit/webalk/library/exceptions/BookExistsException.java
@@ -1,7 +1,7 @@
 package hu.uni.miskolc.iit.webalk.library.exceptions;
 
 public class BookExistsException extends Exception {
-    public BookExistsException(int isbn) {
+    public BookExistsException(long isbn) {
         super("Book already exists with " + isbn + " ISBN number");
     }
 }
